Chore: Rename shared to debugging-protocol-common
Fix #520
